+++ src/main/java/com/avp/common/block_item/AVPBlockItems.java
@@ -2,12 +2,10 @@ package com.avp.common.block_item;
 
 import mod.azure.azurelib.rewrite.animation.cache.AzIdentityRegistry;
 import net.minecraft.core.Registry;
-import net.minecraft.core.component.DataComponents;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.DyeColor;
 import net.minecraft.world.item.Item;
-import net.minecraft.world.item.component.ItemContainerContents;
 import net.minecraft.world.level.block.Block;
 
 import java.util.Arrays;
@@ -124,14 +122,7 @@ public class AVPBlockItems {
 
     public static final BlockItem LEAD_BLOCK = register(AVPBlocks.LEAD_BLOCK);
 
-    public static final BlockItem LEAD_CHEST = register(
-        new Item.Properties().component(DataComponents.CONTAINER, ItemContainerContents.EMPTY)
-            .component(
-                DataComponents.MAX_STACK_SIZE,
-                1
-            ),
-        AVPBlocks.LEAD_CHEST
-    );
+    public static final BlockItem LEAD_CHEST = register("lead_chest", LeadChestItem::new);
 
     public static final BlockItem AMMO_CHEST = register("ammo_chest", AmmoChestItem::new);