+++ src/main/java/com/avp/data/lang/EnglishLanguageProvider.java
@@ -624,6 +624,9 @@ public class EnglishLanguageProvider extends FabricLanguageProvider {
         translationBuilder.add("tooltip.avp.ammo_chest.placed", "When Placed:");
         translationBuilder.add("tooltip.avp.ammo_chest.turret_load_from_chest", "+ Nearby Turrets use Ammo from Chest");
 
+        translationBuilder.add("tooltip.avp.lead_chest.in_inventory", "When In Inventory:");
+        translationBuilder.add("tooltip.avp.lead_chest.auto_store_irradiated_items", "+ Irradiated Items Auto-Stored in Chest");
+
         // Keybinds
         translationBuilder.add("key.avp.reload", "Reload");
         translationBuilder.add("keybind.category.avp.weapons", "AVP Weapons");